Rename subsitutions packages to graal to be consistent with other modules
