+++ extensions/jdbc/jdbc-h2/jdbc-h2-runtime/src/main/java/org/jboss/shamrock/jdbc/h2/runtime/graal/Engine.java
@@ -1,4 +1,4 @@
-package org.jboss.shamrock.jdbc.h2.runtime.graalsubstitutions;
+package org.jboss.shamrock.jdbc.h2.runtime.graal;
 
 import org.h2.engine.ConnectionInfo;
 import org.h2.engine.Session;
